provider_configuration init of MODEL_PLATFORMS
