+++ chatchat-server/tests/unit_server/test_init_server.py
@@ -1,0 +1,6 @@
+from chatchat.model_loaders.init_server import init_server
+
+
+def test_init_server():
+
+    init_server()
